% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add External Resources}
\usage{
add_external_resources()
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} with resource tags
}
\description{
Add CSS, JavaScript, and other static resources to the app.
}
\keyword{internal}
