% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{advance_rotation}
\alias{advance_rotation}
\title{Advance Rotation}
\usage{
advance_rotation(chore_id)
}
\arguments{
\item{chore_id}{The chore ID.}
}
\value{
TRUE on success.
}
\description{
Moves to the next person in the rotation.
}
