% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agenda_view.R
\name{agenda_event_card}
\alias{agenda_event_card}
\title{Create Agenda Event Card}
\usage{
agenda_event_card(event)
}
\arguments{
\item{event}{A single row from the events data frame}
}
\value{
HTML for an agenda event card
}
\description{
Create Agenda Event Card
}
\keyword{internal}
