% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Main Application Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments}
}
\description{
The server-side logic for the Skylight calendar app.
}
\keyword{internal}
