% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{app_sys}
\alias{app_sys}
\title{Get Path to Package Files}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Path components relative to \verb{inst/}}
}
\value{
Character string with the full path
}
\description{
Helper function to access files in the installed package.
}
\keyword{internal}
