% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Main Application UI}
\usage{
app_ui()
}
\value{
A \code{\link[bslib:page_navbar]{bslib::page_navbar()}} UI definition
}
\description{
Constructs the main UI for the Skylight calendar app using bslib.
}
\keyword{internal}
