% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calendar.R
\name{cache_calendars}
\alias{cache_calendars}
\title{Cache Calendars}
\usage{
cache_calendars(calendars)
}
\arguments{
\item{calendars}{Data frame of calendars}
}
\description{
Cache Calendars
}
\keyword{internal}
