% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calendar.R
\name{cache_events}
\alias{cache_events}
\title{Cache Events}
\usage{
cache_events(events)
}
\arguments{
\item{events}{Data frame of events}
}
\description{
Stores events in the local DuckDB cache.
}
\keyword{internal}
