% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_auth.R
\name{calendar_deauth}
\alias{calendar_deauth}
\title{Clear Google Calendar Authentication}
\usage{
calendar_deauth()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Removes cached OAuth tokens and clears the current session.
}
\examples{
calendar_deauth()
}
