% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countdown.R
\name{countdown_form}
\alias{countdown_form}
\title{Countdown Form}
\usage{
countdown_form(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\description{
Form for adding a new countdown event.
}
\keyword{internal}
