% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_cleanup}
\alias{db_cleanup}
\title{Clear Old Data}
\usage{
db_cleanup(days = 30)
}
\arguments{
\item{days}{Number of days to keep. Default is 30.}
}
\value{
Number of deleted rows.
}
\description{
Removes old cached data from the database.
}
\keyword{internal}
