% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_execute}
\alias{db_execute}
\title{Execute Database Statement}
\usage{
db_execute(statement, params = NULL)
}
\arguments{
\item{statement}{SQL statement string.}

\item{params}{Optional list of statement parameters.}
}
\value{
Number of affected rows.
}
\description{
A convenience wrapper for executing SQL statements (INSERT, UPDATE, DELETE).
}
\keyword{internal}
