% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_get_setting}
\alias{db_get_setting}
\title{Get Setting}
\usage{
db_get_setting(key, default = NULL)
}
\arguments{
\item{key}{Setting key.}

\item{default}{Default value if setting doesn't exist.}
}
\value{
The setting value, or the default.
}
\description{
Retrieves a setting value from the database.
}
\keyword{internal}
