% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_save_setting}
\alias{db_save_setting}
\title{Save Setting}
\usage{
db_save_setting(key, value)
}
\arguments{
\item{key}{Setting key.}

\item{value}{Setting value (will be converted to JSON).}
}
\description{
Stores a setting value in the database.
}
\keyword{internal}
