% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{delete_chore}
\alias{delete_chore}
\title{Delete Chore}
\usage{
delete_chore(id)
}
\arguments{
\item{id}{The chore ID.}
}
\value{
TRUE on success.
}
\description{
Soft-deletes a chore by setting is_active to FALSE.
}
