% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{delete_family_member}
\alias{delete_family_member}
\title{Delete Family Member}
\usage{
delete_family_member(id)
}
\arguments{
\item{id}{The member ID.}
}
\value{
TRUE on success.
}
\description{
Soft-deletes a family member by setting is_active to FALSE.
}
