% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_week_view.R
\name{event_card}
\alias{event_card}
\title{Create Event Card UI}
\usage{
event_card(event)
}
\arguments{
\item{event}{A single row from the events data frame}
}
\value{
HTML for an event card
}
\description{
Create Event Card UI
}
\keyword{internal}
