% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_day_view.R
\name{event_card_detailed}
\alias{event_card_detailed}
\title{Create Detailed Event Card}
\usage{
event_card_detailed(event)
}
\arguments{
\item{event}{A single row from the events data frame}
}
\value{
HTML for a detailed event card
}
\description{
Create Detailed Event Card
}
\keyword{internal}
