% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_day_view.R
\name{event_card_simple}
\alias{event_card_simple}
\title{Create Simple Event Card}
\usage{
event_card_simple(event)
}
\arguments{
\item{event}{A single row from the events data frame}
}
\value{
HTML for a simple event card
}
\description{
Create Simple Event Card
}
\keyword{internal}
