% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calendar.R
\name{fetch_calendar_events}
\alias{fetch_calendar_events}
\title{Fetch Events from a Single Calendar}
\usage{
fetch_calendar_events(token, calendar_id, time_min, time_max)
}
\arguments{
\item{token}{OAuth token}

\item{calendar_id}{Calendar ID}

\item{time_min}{Start time (ISO 8601)}

\item{time_max}{End time (ISO 8601)}
}
\value{
Data frame of events
}
\description{
Fetch Events from a Single Calendar
}
\keyword{internal}
