% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_demo.R
\name{generate_sample_calendars}
\alias{generate_sample_calendars}
\title{Generate Sample Calendars for Demo Mode}
\usage{
generate_sample_calendars()
}
\value{
A data frame of sample calendars.
}
\description{
Generate Sample Calendars for Demo Mode
}
\keyword{internal}
