% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_auth.R
\name{get_auth_email}
\alias{get_auth_email}
\title{Get Current User Email}
\usage{
get_auth_email()
}
\value{
Character string with the email address.
}
\description{
Retrieves the email address of the authenticated user.
}
\keyword{internal}
