% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offline.R
\name{get_cache_age}
\alias{get_cache_age}
\title{Get Cache Age in Minutes}
\usage{
get_cache_age()
}
\value{
Numeric. Cache age in minutes, or NULL if no cache.
}
\description{
Get Cache Age in Minutes
}
\keyword{internal}
