% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calendar.R
\name{get_cached_calendars}
\alias{get_cached_calendars}
\title{Get Cached Calendars}
\usage{
get_cached_calendars()
}
\value{
Data frame of cached calendars, or NULL.
}
\description{
Get Cached Calendars
}
\keyword{internal}
