% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{get_chat_response}
\alias{get_chat_response}
\title{Get Chat Response}
\usage{
get_chat_response(chat, message, events, selected_date)
}
\arguments{
\item{chat}{An ellmer chat object}

\item{message}{User's message}

\item{events}{Current events data}

\item{selected_date}{Current selected date}
}
\value{
Character string with the response
}
\description{
Send a message to the LLM and get a response.
}
\keyword{internal}
