% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{get_chore}
\alias{get_chore}
\title{Get Chore by ID}
\usage{
get_chore(id)
}
\arguments{
\item{id}{The chore ID.}
}
\value{
A single-row data frame or NULL if not found.
}
\description{
Get Chore by ID
}
