% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_countdown.R
\name{get_countdown_summary}
\alias{get_countdown_summary}
\title{Get Countdown Summary}
\usage{
get_countdown_summary()
}
\value{
A list with next countdown info, or NULL.
}
\description{
Returns a summary for displaying in the widget.
}
