% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_dark_theme}
\alias{get_dark_theme}
\title{Dark Mode Theme}
\usage{
get_dark_theme()
}
\value{
A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object
}
\description{
Creates a dark variant of the app theme.
}
\keyword{internal}
