% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{get_db_path}
\alias{get_db_path}
\title{Get Database Path}
\usage{
get_db_path()
}
\value{
Character string with the database path.
}
\description{
Returns the path to the DuckDB database file.
}
\keyword{internal}
