% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{get_family_member}
\alias{get_family_member}
\title{Get Family Member by ID}
\usage{
get_family_member(id)
}
\arguments{
\item{id}{The member ID.}
}
\value{
A single-row data frame or NULL if not found.
}
\description{
Get Family Member by ID
}
