% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamification.R
\name{get_longest_streak}
\alias{get_longest_streak}
\title{Get Longest Streak}
\usage{
get_longest_streak(member_id)
}
\arguments{
\item{member_id}{The member ID.}
}
\value{
Number of days in longest streak (integer).
}
\description{
Calculates the longest completion streak ever achieved by a member.
}
