% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_countdown.R
\name{get_next_countdown}
\alias{get_next_countdown}
\title{Get Next Countdown}
\usage{
get_next_countdown()
}
\value{
A single-row data frame, or NULL if none.
}
\description{
Returns the next upcoming countdown event.
}
