% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamification.R
\name{get_period_points}
\alias{get_period_points}
\title{Get Period Points}
\usage{
get_period_points(member_id, period = "week")
}
\arguments{
\item{member_id}{The member ID.}

\item{period}{One of 'day', 'week', 'month', 'all_time'.}
}
\value{
Total points (integer).
}
\description{
Gets total points earned by a member in a time period.
}
