% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{get_rotation_assignee}
\alias{get_rotation_assignee}
\title{Get Rotation Assignee}
\usage{
get_rotation_assignee(chore_id)
}
\arguments{
\item{chore_id}{The chore ID.}
}
\value{
The member ID who should be assigned, or NULL.
}
\description{
Gets the current assignee based on rotation.
}
\keyword{internal}
