% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{get_rotation_config}
\alias{get_rotation_config}
\title{Get Rotation Config}
\usage{
get_rotation_config(chore_id)
}
\arguments{
\item{chore_id}{The chore ID.}
}
\value{
Rotation config data frame or NULL.
}
\description{
Get Rotation Config
}
\keyword{internal}
