% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{get_theme}
\alias{get_theme}
\title{Build App Theme from brand.yml}
\usage{
get_theme()
}
\value{
A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object
}
\description{
Creates a bslib theme based on brand.yml configuration.
}
\keyword{internal}
