% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamification.R
\name{get_today_points}
\alias{get_today_points}
\title{Get Today's Points}
\usage{
get_today_points(member_id)
}
\arguments{
\item{member_id}{The member ID.}
}
\value{
Total points (integer).
}
\description{
Gets total points earned by a member today.
}
