% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_auth.R
\name{get_token_cache_dir}
\alias{get_token_cache_dir}
\title{Get Token Cache Directory}
\usage{
get_token_cache_dir()
}
\value{
Path to the token cache directory
}
\description{
Get Token Cache Directory
}
\keyword{internal}
