% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{get_week_dates}
\alias{get_week_dates}
\title{Get Week Dates}
\usage{
get_week_dates(date, week_start = 0)
}
\arguments{
\item{date}{A Date object.}

\item{week_start}{Day the week starts: 0 = Sunday, 1 = Monday.}
}
\value{
A vector of 7 Date objects.
}
\description{
Returns all dates for the week containing the given date.
}
\keyword{internal}
