% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{is_current_month}
\alias{is_current_month}
\title{Check if Date is in Current Month}
\usage{
is_current_month(date, reference = Sys.Date())
}
\arguments{
\item{date}{A Date object.}

\item{reference}{Reference date for comparison.}
}
\value{
Logical.
}
\description{
Check if Date is in Current Month
}
\keyword{internal}
