% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auto_dark_mode.R
\name{is_dark_time}
\alias{is_dark_time}
\title{Check if Current Time is in Dark Period}
\usage{
is_dark_time(sunrise_time, sunset_time)
}
\arguments{
\item{sunrise_time}{Character. Sunrise time in "HH:MM" format.}

\item{sunset_time}{Character. Sunset time in "HH:MM" format.}
}
\value{
Logical. TRUE if current time is in dark period.
}
\description{
Check if Current Time is in Dark Period
}
\keyword{internal}
