% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{is_today}
\alias{is_today}
\title{Check if Date is Today}
\usage{
is_today(date)
}
\arguments{
\item{date}{A Date object.}
}
\value{
Logical.
}
\description{
Check if Date is Today
}
\keyword{internal}
