% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_holidays.R
\name{is_us_holiday}
\alias{is_us_holiday}
\title{Check if Date is a US Holiday}
\usage{
is_us_holiday(date)
}
\arguments{
\item{date}{A Date object.}
}
\value{
Logical indicating if the date is a US federal holiday.
}
\description{
Check if Date is a US Holiday
}
\keyword{internal}
