% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{is_weekend}
\alias{is_weekend}
\title{Check if Date is Weekend}
\usage{
is_weekend(date)
}
\arguments{
\item{date}{A Date object.}
}
\value{
Logical.
}
\description{
Check if Date is Weekend
}
\keyword{internal}
