% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auto_dark_mode.R
\name{load_auto_dark_settings}
\alias{load_auto_dark_settings}
\title{Load Auto Dark Mode Settings}
\usage{
load_auto_dark_settings()
}
\value{
List with auto dark mode settings.
}
\description{
Load Auto Dark Mode Settings
}
\keyword{internal}
