% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_recurrence.R
\name{matches_weekly}
\alias{matches_weekly}
\title{Check Weekly Recurrence Match}
\usage{
matches_weekly(date, rule, start_date, interval)
}
\description{
Check Weekly Recurrence Match
}
\keyword{internal}
