% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_family.R
\name{member_form}
\alias{member_form}
\title{Member Form}
\usage{
member_form(ns, member = NULL)
}
\arguments{
\item{ns}{Namespace function.}

\item{member}{Optional existing member data for editing.}
}
\value{
HTML form elements
}
\description{
Creates the form fields for adding/editing a member.
}
\keyword{internal}
