% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_family.R
\name{member_row}
\alias{member_row}
\title{Member Row UI}
\usage{
member_row(member, ns)
}
\arguments{
\item{member}{A single-row data frame with member info.}

\item{ns}{Namespace function.}
}
\value{
HTML div
}
\description{
Creates a single member row for the list.
}
\keyword{internal}
