% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agenda_view.R
\name{mod_agenda_view_server}
\alias{mod_agenda_view_server}
\title{Agenda View Module Server}
\usage{
mod_agenda_view_server(id, events, selected_date)
}
\arguments{
\item{id}{Module namespace ID}

\item{events}{Reactive containing event data}

\item{selected_date}{Reactive value for selected date}
}
\description{
Agenda View Module Server
}
\keyword{internal}
