% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agenda_view.R
\name{mod_agenda_view_ui}
\alias{mod_agenda_view_ui}
\title{Agenda View Module UI}
\usage{
mod_agenda_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
Agenda View Module UI
}
\keyword{internal}
