% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auth.R
\name{mod_auth_server}
\alias{mod_auth_server}
\title{Authentication Module Server}
\usage{
mod_auth_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A reactive indicating authentication status (TRUE/FALSE)
}
\description{
Authentication Module Server
}
\keyword{internal}
