% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auth.R
\name{mod_auth_ui}
\alias{mod_auth_ui}
\title{Authentication Module UI}
\usage{
mod_auth_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
UI for Google Calendar authentication status and controls.
}
\keyword{internal}
