% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auto_dark_mode.R
\name{mod_auto_dark_mode_ui}
\alias{mod_auto_dark_mode_ui}
\title{Auto Dark Mode Module UI}
\usage{
mod_auto_dark_mode_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Settings UI for automatic dark mode based on time of day.
}
\keyword{internal}
