% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{mod_chat_server}
\alias{mod_chat_server}
\title{Chat Module Server}
\usage{
mod_chat_server(id, events, calendars, selected_date)
}
\arguments{
\item{id}{Module namespace ID}

\item{events}{Reactive containing event data}

\item{calendars}{Reactive containing calendar list}

\item{selected_date}{Reactive value for selected date}
}
\description{
Chat Module Server
}
\keyword{internal}
