% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chores.R
\name{mod_chores_server}
\alias{mod_chores_server}
\title{Chores Dashboard Module Server}
\usage{
mod_chores_server(id, family_members = NULL)
}
\arguments{
\item{id}{Module namespace ID}

\item{family_members}{Reactive containing family members list (optional)}
}
\description{
Chores Dashboard Module Server
}
\keyword{internal}
