% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chores.R
\name{mod_chores_ui}
\alias{mod_chores_ui}
\title{Chores Dashboard Module UI}
\usage{
mod_chores_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
Main chores dashboard with multiple view modes.
}
\keyword{internal}
