% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clock.R
\name{mod_clock_ui}
\alias{mod_clock_ui}
\title{Clock Module UI}
\usage{
mod_clock_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
A live-updating clock display for the navbar.
}
\keyword{internal}
