% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countdown.R
\name{mod_countdown_server}
\alias{mod_countdown_server}
\title{Countdown Widget Module Server}
\usage{
mod_countdown_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Displays the next upcoming countdown in a compact format.
}
\keyword{internal}
