% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countdown.R
\name{mod_countdown_ui}
\alias{mod_countdown_ui}
\title{Countdown Widget Module UI}
\usage{
mod_countdown_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
A compact countdown display for the navbar showing days until next event.
Clicking opens a modal to manage countdowns.
}
\keyword{internal}
