% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_day_view.R
\name{mod_day_view_ui}
\alias{mod_day_view_ui}
\title{Day View Module UI}
\usage{
mod_day_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
Day View Module UI
}
\keyword{internal}
