% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_family.R
\name{mod_family_server}
\alias{mod_family_server}
\title{Family Members Module Server}
\usage{
mod_family_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A reactive returning the list of family members
}
\description{
Family Members Module Server
}
\keyword{internal}
