% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_family.R
\name{mod_family_ui}
\alias{mod_family_ui}
\title{Family Members Module UI}
\usage{
mod_family_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
UI for managing family members in the chores system.
}
\keyword{internal}
