% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_month_view.R
\name{mod_month_view_server}
\alias{mod_month_view_server}
\title{Month View Module Server}
\usage{
mod_month_view_server(id, events, selected_date, calendars)
}
\arguments{
\item{id}{Module namespace ID}

\item{events}{Reactive containing event data}

\item{selected_date}{Reactive value for selected date}

\item{calendars}{Reactive containing calendar list}
}
\description{
Month View Module Server
}
\keyword{internal}
