% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_month_view.R
\name{mod_month_view_ui}
\alias{mod_month_view_ui}
\title{Month View Module UI}
\usage{
mod_month_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
Month View Module UI
}
\keyword{internal}
