% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_weather.R
\name{mod_weather_server}
\alias{mod_weather_server}
\title{Weather Widget Module Server}
\usage{
mod_weather_server(id, lat = NULL, lon = NULL)
}
\arguments{
\item{id}{Module namespace ID}

\item{lat}{Reactive or static latitude (default: NULL, will use config)}

\item{lon}{Reactive or static longitude (default: NULL, will use config)}
}
\description{
Fetches and displays weather data from OpenWeatherMap API.
}
\keyword{internal}
