% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_week_view.R
\name{mod_week_view_server}
\alias{mod_week_view_server}
\title{Week View Module Server}
\usage{
mod_week_view_server(id, events, selected_date, calendars)
}
\arguments{
\item{id}{Module namespace ID}

\item{events}{Reactive containing event data}

\item{selected_date}{Reactive value for selected date}

\item{calendars}{Reactive containing calendar list}
}
\description{
Week View Module Server
}
\keyword{internal}
