% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_week_view.R
\name{mod_week_view_ui}
\alias{mod_week_view_ui}
\title{Week View Module UI}
\usage{
mod_week_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI definition
}
\description{
Week View Module UI
}
\keyword{internal}
