% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offline.R
\docType{data}
\name{offline_env}
\alias{offline_env}
\title{Package Environment for Offline State}
\format{
An object of class \code{environment} of length 3.
}
\usage{
offline_env
}
\description{
Package Environment for Offline State
}
\keyword{internal}
