% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offline.R
\name{offline_indicator_ui}
\alias{offline_indicator_ui}
\title{Create Offline Indicator UI}
\usage{
offline_indicator_ui(id = "offline")
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
Shiny UI elements.
}
\description{
Returns UI elements for displaying offline status.
}
\keyword{internal}
