% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auto_dark_mode.R
\name{parse_time_to_minutes}
\alias{parse_time_to_minutes}
\title{Parse Time String to Minutes Since Midnight}
\usage{
parse_time_to_minutes(time_str)
}
\arguments{
\item{time_str}{Character. Time in "HH:MM" format.}
}
\value{
Numeric. Minutes since midnight.
}
\description{
Parse Time String to Minutes Since Midnight
}
\keyword{internal}
