% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_auth.R
\name{refresh_token}
\alias{refresh_token}
\title{Refresh OAuth Token}
\usage{
refresh_token(token)
}
\arguments{
\item{token}{An OAuth token object}
}
\value{
A refreshed OAuth token object
}
\description{
Refresh OAuth Token
}
\keyword{internal}
