% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_countdown.R
\name{remove_countdown}
\alias{remove_countdown}
\title{Remove Countdown Event}
\usage{
remove_countdown(id)
}
\arguments{
\item{id}{The countdown ID to remove.}
}
\value{
TRUE on success.
}
\description{
Removes an event from the countdown tracker.
}
