% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{remove_rotation}
\alias{remove_rotation}
\title{Remove Rotation}
\usage{
remove_rotation(chore_id)
}
\arguments{
\item{chore_id}{The chore ID.}
}
\value{
TRUE on success.
}
\description{
Deactivates rotation for a chore.
}
