% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{save_chat_message}
\alias{save_chat_message}
\title{Save Chat Message to Database}
\usage{
save_chat_message(user_message, assistant_response)
}
\arguments{
\item{user_message}{The user's message}

\item{assistant_response}{The assistant's response}
}
\description{
Save Chat Message to Database
}
\keyword{internal}
