% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offline.R
\name{set_online_status}
\alias{set_online_status}
\title{Set Online/Offline Status}
\usage{
set_online_status(is_online)
}
\arguments{
\item{is_online}{Logical. TRUE if online.}
}
\description{
Updates the cached connectivity status.
}
\keyword{internal}
