% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{skip_assignment}
\alias{skip_assignment}
\title{Skip Assignment}
\usage{
skip_assignment(assignment_id)
}
\arguments{
\item{assignment_id}{The assignment ID.}
}
\value{
TRUE on success.
}
\description{
Marks an assignment as skipped.
}
