% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skylight_rcpp}
\alias{skylight_rcpp}
\title{fast C++ implementation of skylight}
\usage{
skylight_rcpp(forcing)
}
\arguments{
\item{forcing}{input matrix with forcing parameters organized
as longitude, latitude, year, month, day, hour, minutes,
sky_conditions}
}
\value{
sky illuminance results as a matrix
}
\description{
Calculates sky illuminance values faster. This function
should not be called independently and the formal R
skylight() function should be used with the parameter
fast = TRUE!
}
