% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamification.R
\docType{data}
\name{STREAK_BONUSES}
\alias{STREAK_BONUSES}
\title{Streak bonus thresholds}
\format{
An object of class \code{list} of length 5.
}
\usage{
STREAK_BONUSES
}
\description{
Points awarded for maintaining completion streaks.
}
\keyword{internal}
