% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{update_chore}
\alias{update_chore}
\title{Update Chore}
\usage{
update_chore(id, ...)
}
\arguments{
\item{id}{The chore ID.}

\item{...}{Fields to update.}
}
\value{
TRUE on success.
}
\description{
Update Chore
}
