% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{update_family_member}
\alias{update_family_member}
\title{Update Family Member}
\usage{
update_family_member(id, ...)
}
\arguments{
\item{id}{The member ID.}

\item{...}{Fields to update (name, display_name, avatar_emoji, color, birth_date, is_active).}
}
\value{
TRUE on success.
}
\description{
Update Family Member
}
