% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{validate_enum}
\alias{validate_enum}
\title{Validate Enum Value}
\usage{
validate_enum(value, name, allowed)
}
\arguments{
\item{value}{The value to validate.}

\item{name}{The parameter name for error messages.}

\item{allowed}{Vector of allowed values.}
}
\value{
The value if valid.
}
\description{
Validate Enum Value
}
\keyword{internal}
