% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{should_run_today}
\alias{should_run_today}
\title{Should Chore Run Today}
\usage{
should_run_today(chore, date)
}
\arguments{
\item{chore}{A single-row chore data frame.}

\item{date}{The date to check.}
}
\value{
TRUE if the chore should run, FALSE otherwise.
}
\description{
Determines if a chore should be assigned on a given date based on frequency
or recurrence rule.
}
\keyword{internal}
