% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chores.R
\name{uncomplete_assignment}
\alias{uncomplete_assignment}
\title{Uncomplete Assignment}
\usage{
uncomplete_assignment(assignment_id)
}
\arguments{
\item{assignment_id}{The assignment ID.}
}
\value{
TRUE on success.
}
\description{
Reverts a completed assignment back to pending.
}
